requirements.txt
torch
torcheval
wandb
matplotlib
opencv-python
tqdm
einops
pyyaml
path
numpy
